--- test/mako/alerter/slack/config.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2019 The Knative Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package slack
-
-import (
-	yaml "gopkg.in/yaml.v2"
-	configPkg "knative.dev/pkg/test/mako/config"
-)
-
-// Config contains all repo configs for performance regression alerting.
-type Config struct {
-	Channels []Channel `yaml:"channels"`
-}
-
-// Channel contains Slack channel's info
-type Channel struct {
-	Name     string `yaml:"name"`
-	Identity string `yaml:"identity"`
-}
-
-// loadConfig parses config from configFile
-func loadConfig() ([]Channel, error) {
-	content, err := configPkg.LoadSlackConfig()
-	if err != nil {
-		return nil, err
-	}
-	return parseConfig(content)
-}
-
-func parseConfig(configYaml []byte) ([]Channel, error) {
-	conf := &Config{}
-	if err := yaml.Unmarshal(configYaml, conf); err != nil {
-		return nil, err
-	}
-	return conf.Channels, nil
-}
