+++ test/mako/alerter/slack/message.go
@@ -23,6 +23,7 @@ import (
 	"time"
 
 	"knative.dev/pkg/test/helpers"
+	"knative.dev/pkg/test/mako/config"
 	"knative.dev/pkg/test/slackutil"
 )
 
@@ -38,13 +39,13 @@ As of %s, there is a new performance regression detected from automation test:
 type MessageHandler struct {
 	readClient  slackutil.ReadOperations
 	writeClient slackutil.WriteOperations
-	channels    []Channel
+	channels    []config.Channel
 	dryrun      bool
 }
 
 // Setup creates the necessary setup to make calls to work with slack
 func Setup(userName, readTokenPath, writeTokenPath, repo string, dryrun bool) (*MessageHandler, error) {
-	channels, err := loadConfig()
+	channels, err := config.LoadSlackConfig()
 	if err != nil {
 		return nil, fmt.Errorf("failed to load the slack config: %v", err)
 	}
