+++ test/mako/config/testdata/config-slack.yaml
